# app/schemas/model_schemas.py
"""
This module defines the Pydantic models (schemas) for the API.
"""

from datetime import datetime
from typing import Optional

from pydantic import BaseModel, ConfigDict, EmailStr, Field

from app.services.visibility import VisibilityTier


# User data provided as input
class InUserBase(BaseModel):
    """
    Schema for user input data.
    """

    display_name: str
    email: str
    password: str


# User data returned in a response
class OutUserDisplay(BaseModel):
    """
    Schema for displaying user data.
    """

    id: str
    display_name: str
    email: EmailStr
    user_group: str
    created_at: datetime
    updated_at: datetime

    model_config = ConfigDict(
        arbitrary_types_allowed=True,
        from_attributes=True,
    )


class OutResponseBaseModel(BaseModel):
    """
    All responses will have this format.
    """

    success: bool = True
    data: Optional[list] = []
    message: Optional[str] = 'None'

    model_config = ConfigDict(
        arbitrary_types_allowed=True,
        from_attributes=True,
        # These appear not to work. Ideally we would exclude None and unset.
        exclude_none=True,
        exclude_unset=True,
    )


class OutResponseUserModel(OutResponseBaseModel):
    """
    Response format for user data.
    """

    data: list[OutUserDisplay]

    model_config = ConfigDict(
        arbitrary_types_allowed=True,
        from_attributes=True,
    )


class InApiKeyCreate(BaseModel):
    """
    Request body for minting an API key.
    """

    name: str = Field(min_length=1, max_length=60)


class OutApiKey(BaseModel):
    """
    An API key as shown in listings — never includes the secret.
    """

    id: str
    name: str
    prefix: str
    created_at: datetime
    last_used_at: Optional[datetime] = None

    model_config = ConfigDict(from_attributes=True)


class OutApiKeyCreated(OutApiKey):
    """
    Creation response: the one and only time the plaintext key appears.
    """

    key: str


class InRefreshToken(BaseModel):
    """
    Request body for the refresh and sign-out endpoints (#246).
    """

    refresh_token: str = Field(min_length=1)


class OutToken(BaseModel):
    """
    What every sign-in and refresh returns.

    Both lifetimes are reported in seconds so a client can size its cookies
    off the response instead of hardcoding guesses that drift from
    ``ACCESS_TOKEN_EXPIRE_MINUTES`` and ``REFRESH_TOKEN_EXPIRE_DAYS`` — which
    is exactly how the web session cookie ended up outliving its token.
    """

    access_token: str
    refresh_token: str
    token_type: str = 'bearer'
    expires_in: int
    refresh_expires_in: int
    user_id: str
    user_group: str
    email: str


class InVisibilityUpdate(BaseModel):
    """
    Request body for visibility settings (#274). Every setting is a tier —
    ``private``, ``friends`` or ``public`` — not a boolean.

    Only sent fields change; a null handle clears it (allowed only while
    everything is private). An unsent field and an explicit null are both
    left alone, so a client can PUT one radio button without echoing the
    other eight.
    """

    handle: Optional[str] = None
    visibility_profile: Optional[VisibilityTier] = None
    visibility_movies: Optional[VisibilityTier] = None
    visibility_tv: Optional[VisibilityTier] = None
    visibility_books: Optional[VisibilityTier] = None
    visibility_games: Optional[VisibilityTier] = None
    visibility_watchlist_movies: Optional[VisibilityTier] = None
    visibility_watchlist_tv: Optional[VisibilityTier] = None
    visibility_watchlist_books: Optional[VisibilityTier] = None
    visibility_watchlist_games: Optional[VisibilityTier] = None


class OutVisibility(BaseModel):
    """
    The caller's visibility settings, one tier per setting.
    """

    handle: Optional[str] = None
    visibility_profile: VisibilityTier = VisibilityTier.PRIVATE
    visibility_movies: VisibilityTier = VisibilityTier.PRIVATE
    visibility_tv: VisibilityTier = VisibilityTier.PRIVATE
    visibility_books: VisibilityTier = VisibilityTier.PRIVATE
    visibility_games: VisibilityTier = VisibilityTier.PRIVATE
    visibility_watchlist_movies: VisibilityTier = VisibilityTier.PRIVATE
    visibility_watchlist_tv: VisibilityTier = VisibilityTier.PRIVATE
    visibility_watchlist_books: VisibilityTier = VisibilityTier.PRIVATE
    visibility_watchlist_games: VisibilityTier = VisibilityTier.PRIVATE

    model_config = ConfigDict(from_attributes=True)


class InFriendRequest(BaseModel):
    """
    Request body for sending a friend request (#275).

    A handle and nothing else: there is no endpoint that lists or searches
    users, so an exact handle someone told you is the only way to reach them.
    """

    handle: str = Field(min_length=1, max_length=30)


class OutFriendAck(BaseModel):
    """
    Deliberately contentless acknowledgement of a friend-graph write.

    Send and decline both answer with this, and send answers with *exactly*
    this whether or not the handle resolved to anybody — see
    ``router_friends`` for why the response must not vary.
    """

    message: str


class OutFriendUser(BaseModel):
    """
    The other party in a friendship, as much of them as a friend may see.

    No email, no visibility settings — a friendship is not an introduction to
    someone's account.
    """

    id: str
    handle: Optional[str] = None
    display_name: Optional[str] = None

    model_config = ConfigDict(from_attributes=True)


class OutFriend(BaseModel):
    """One accepted friendship. ``id`` addresses the friendship, not the user."""

    id: str
    user: OutFriendUser
    friends_since: datetime


class OutFriendRequest(BaseModel):
    """One pending request, from the perspective of whoever is listing it."""

    id: str
    user: OutFriendUser
    requested_at: datetime


class OutPendingFriendRequests(BaseModel):
    """
    Both directions in one response.

    The inbox badge and the "pending" state of an outgoing request are drawn
    together, so splitting these across two endpoints would only ever mean
    two calls.
    """

    incoming: list[OutFriendRequest] = []
    outgoing: list[OutFriendRequest] = []


class OutFollowUser(BaseModel):
    """
    The other party in a follow, as much of them as a follower may see.

    A separate class from ``OutFriendUser`` even though the fields match: the
    two relationships are deliberately kept apart everywhere else (#276), and
    a schema shared between them would be the one place that split quietly
    disappeared.
    """

    id: str
    handle: Optional[str] = None
    display_name: Optional[str] = None

    model_config = ConfigDict(from_attributes=True)


class OutFollow(BaseModel):
    """
    One follow edge, from whichever side is listing it.

    ``id`` addresses the follow row, not either user. ``followed_at`` is the
    one fact both sides are shown; there is nothing else to negotiate since a
    follow needs no approval.
    """

    id: str
    user: OutFollowUser
    followed_at: datetime


class OutSummaryEntry(BaseModel):
    """One ranked entry on a shelf's Top 5."""

    rank: int
    id: str
    title: str
    year: Optional[int] = None
    poster_url: Optional[str] = None


class OutSummaryShelf(BaseModel):
    """One domain's headline numbers plus its best-ranked entries."""

    category: str
    label: str
    ranked_count: int
    queued_count: int
    public: bool
    top: list[OutSummaryEntry]


class OutSummary(BaseModel):
    """
    Everything the home page renders, in one bounded response — see
    app/services/summary.py for why this endpoint exists.
    """

    handle: Optional[str] = None
    display_name: Optional[str] = None
    profile_public: bool = False
    shelves: list[OutSummaryShelf]
    total_ranked: int