# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "druthers.io API local",
    "description": "API for druthers.io \u2014 track and rank the movies, TV, books, and games you actually care about.",
    "contact": {
      "name": "Druthers",
      "url": "https://www.druthers.io/"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://api.druthers.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/v1/users/me/following/{handle}": {
      "put": {
        "tags": [
          "Follows"
        ],
        "summary": "Follow User",
        "description": "Follow a public profile. Idempotent: following twice just returns the\nexisting row rather than erroring, since there is no second state for a\nrepeat call to conflict with.",
        "operationId": "follow_user_v1_users_me_following__handle__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutFollow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Follows"
        ],
        "summary": "Unfollow User",
        "description": "Stop following someone.\n\nDeliberately does not require the target to still be public \u2014 a follow\npersists through a followee tightening their profile (that's the whole\npoint of #276's \"grants nothing, persists anyway\" rule), and the follower\nmust still be able to unfollow after that happens.",
        "operationId": "unfollow_user_v1_users_me_following__handle__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/me/following": {
      "get": {
        "tags": [
          "Follows"
        ],
        "summary": "List Following",
        "description": "Everyone the caller follows.",
        "operationId": "list_following_v1_users_me_following_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutFollow"
                  },
                  "type": "array",
                  "title": "Response List Following V1 Users Me Following Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/users/me/followers": {
      "get": {
        "tags": [
          "Follows"
        ],
        "summary": "List Followers",
        "description": "Everyone following the caller.",
        "operationId": "list_followers_v1_users_me_followers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutFollow"
                  },
                  "type": "array",
                  "title": "Response List Followers V1 Users Me Followers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutFollow": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/OutFollowUser"
          },
          "followed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Followed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "followed_at"
        ],
        "title": "OutFollow",
        "description": "One follow edge, from whichever side is listing it.\n\n``id`` addresses the follow row, not either user. ``followed_at`` is the\none fact both sides are shown; there is nothing else to negotiate since a\nfollow needs no approval."
      },
      "OutFollowUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "OutFollowUser",
        "description": "The other party in a follow, as much of them as a follower may see.\n\nA separate class from ``OutFriendUser`` even though the fields match: the\ntwo relationships are deliberately kept apart everywhere else (#276), and\na schema shared between them would be the one place that split quietly\ndisappeared."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  },
...
}